requirements.txt
Django==5.0.7
djangorestframework==3.15.2
firebase-admin==6.5.0
python-dotenv==1.0.1
requests==2.32.3